requirements.txt
requests
urllib3
pillow